Add Parquet validation utility and refactor tests
- Introduced a new utility function `assert_valid_parquet_file` in `common/parquet.rs` to validate Parquet files, ensuring they contain at least one column.
- Refactored existing tests in `cli.rs` and `repl.rs` to utilize the new validation function, improving code clarity and reducing duplication.
- Updated `common/mod.rs` to include the new Parquet module for better organization of common test utilities.
