+++ tests/cli.rs
@@ -13,9 +13,9 @@ mod common;
 
 use common::TEMPDIR_PLACEHOLDER;
 use common::assert_output_contains;
+use common::assert_valid_parquet_file;
 use common::get_row_count;
 use common::replace_tempdir;
-use parquet::file::reader::FileReader as _;
 
 #[derive(Debug, Default, World)]
 pub struct CliWorld {
@@ -358,14 +358,7 @@ fn that_file_should_be_valid_parquet(world: &mut CliWorld) {
         .last_file
         .as_ref()
         .expect("No file has been set; use 'the file \"...\" should exist' first");
-    let file = std::fs::File::open(path).expect("Failed to open file");
-    let reader = parquet::file::reader::SerializedFileReader::new(file)
-        .expect("Expected file to be valid Parquet, but reading failed");
-    let metadata = reader.metadata();
-    assert!(
-        !metadata.file_metadata().schema().get_fields().is_empty(),
-        "Expected Parquet file to have at least one column"
-    );
+    assert_valid_parquet_file(path);
 }
 
 #[then(regex = r#"^the file "(.+)" should contain:$"#)]
