+++ tests/common/mod.rs
@@ -1,7 +1,9 @@
 mod assert_output;
+mod parquet;
 mod row_count;
 
 pub use assert_output::TEMPDIR_PLACEHOLDER;
 pub use assert_output::assert_output_contains;
 pub use assert_output::replace_tempdir;
+pub use parquet::assert_valid_parquet_file;
 pub use row_count::get_row_count;
